% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkPLUS-package.R
\docType{package}
\name{tkPLUS-package}
\alias{tkPLUS}
\alias{tkPLUS-package}
\title{tkPLUS: tkPLUS: A tcltk Wrapper Wrapper}
\description{
A wrapper for the base R Tcl/Tk wrapper tcltk.
}
\keyword{internal}
